pyproject.toml
[tool.poetry]
name = "texas-leg-scraper"
version = "0.1.0"
description = ""
authors = ["lwgover <lucasgover@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.2"
feedparser = "^6.0.11"
requests = "^2.32.3"
pandas = "^2.2.3"
bs4 = "^0.0.2"
pyyaml = "^6.0.2"
ipykernel = "^6.29.5"
gspread = "^6.2.0"
lxml = "^5.3.1"
dbt-core = "^1.9.2"
dbt-duckdb = "^1.9.2"
dlt = {extras = ["duckdb"], version = "^1.8.0"}
streamlit = "^1.43.0"
dotenv = "^0.9.9"
pypdf2 = "^3.0.1"
pdfplumber = "^0.11.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
